Improve "Using BLAS library" message
The previous message was a little confusing when the used lib name was a pattern rather than a filename. With this change we will indicate that the lib name is a "pattern". Additionally, this makes a small change to one of the prefixes (to make it a tiny bit more readable) and to one of the suffixes (to remove a duplicated empty pattern).